restructure components, extract SignOut from Profile, delete commented lines
