+++ src/components/Profile.jsx
@@ -1,12 +1,11 @@
-import { Pressable, View, StyleSheet, Image } from "react-native";
+import { View, StyleSheet, Image } from "react-native";
 import Text from "./Text";
 import Loader from "./Loader";
-import useSignOut from "../hooks/useSignOut";
-import { confirmAction } from "../utils/alert";
 import theme from "../theme";
 import useCurrentUser from "../hooks/useCurrentUser";
 import user from "../../assets/user.png";
 import format from "date-fns/format";
+import SignOut from "./SignOut";
 
 const styles = StyleSheet.create({
   container: {
@@ -25,7 +24,6 @@ const styles = StyleSheet.create({
     width: 150,
     height: 150,
     margin: 10,
-    // backgroundColor: theme.colors.dark,
     borderWidth: 2,
     borderRadius: 75,
     justifyContent: "center",
@@ -39,37 +37,12 @@ const styles = StyleSheet.create({
     alignItems: "center",
     margin: 5,
   },
-  button: {
-    backgroundColor: theme.colors.primary,
-    margin: 5,
-    height: 50,
-    borderRadius: 5,
-    justifyContent: "center",
-    alignItems: "center",
-  },
 });
 
 const Profile = () => {
   const { loading, error, me } = useCurrentUser();
-  const [signOut] = useSignOut();
   if (loading) return <Loader />;
   if (error) return <Text>{error.message}</Text>;
-
-  const onSignOutPressed = async () => {
-    // console.log("Sign out!");
-    await signOut();
-    // navigate("/SignIn");
-  };
-
-  const onSignOut = () => {
-    confirmAction(
-      "Log out",
-      "Are you sure you want to log out",
-      "Yes",
-      "No",
-      onSignOutPressed
-    );
-  };
 
   return (
     <View style={styles.container}>
@@ -98,11 +71,7 @@ const Profile = () => {
           </View>
         </View>
       </View>
-      <Pressable onPress={onSignOut} style={styles.button}>
-        <Text color="white" fontWeight="bold">
-          Sign Out
-        </Text>
-      </Pressable>
+      <SignOut />
     </View>
   );
 };
